Update package name.
