--- src/java/week01/task1/CombineWords_Melike.java
@@ -1,15 +1,0 @@
-package java.week01.task1;
-
-public class CombineWords_Melike {
-    /*
-    Create a method that can combine two Strings then print the combined version.
-    But if the last letter of the first String and the first letter of the second String is the same, print that character once.
-
-    Example;
-                Input:
-                   String1= one
-                   String2 = eight
-                Output:
-                     oneight
-     */
-}
