--- src/java/week01/task2/EmailTask1_Emir.java
@@ -1,15 +1,0 @@
-package java.week01.task2;
-
-public class EmailTask1_Emir {
-    /*
-
-    Create a method that can swap first name with last name in the email (Separated by an underscore).
-    Assume that email address is constructed by person's first name and followed by an underscore and last name.
-    If the email doesn't contain an underscore print the given input email.
-
-    Example;
-
-            input: mike_tyson@gmail.com
-            output: tyson_mike@gmail.com
- */
-}
