--- src/java/week01/task3/EmailTask2_Buse.java
@@ -1,18 +1,0 @@
-package java.week01.task3;
-
-public class EmailTask2_Buse {
-/*
-   Create a method that can print out information about user based on email. Print first name, last name, and domain.
-   Assume that email address is constructed by person's first name and followed by an underscore and last name.
-   First and Last name should be printed with proper format - uppercase first letter and remaining lowercase.
-
-   Example:
-               Input:
-                   craig_federighi@apple.com
-
-               Output:
-                   First name: Craig
-                   Last name: Federighi
-                   Domain: apple
-*/
-}
