--- src/java/week01/task5/ReplaceX_Emir.java
@@ -1,13 +1,0 @@
-package java.week01.task5;
-
-public class ReplaceX_Emir {
-/*
-Create a method that accept a String.And replace all the x or X with character a.
-
-Example;
-                    Input:
-                        xcodeX
-                    Output:
-                        acodea
- */
-}
