--- src/java/week01/task6/StartsWithX_Emir.java
@@ -1,16 +1,0 @@
-package java.week01.task6;
-
-public class StartsWithX_Emir {
-/*
-Create a method that accept a String.
-If the String starts with x, replace it with a.
-
-Example;
-
-                    Input:
-                        xcodex
-                    Output:
-                        acodex
-
- */
-}
