--- src/java/week01/task7/WithoutFirstChar_Buse.java
@@ -1,17 +1,0 @@
-package java.week01.task7;
-
-public class WithoutFirstChar_Buse {
-/*
- Create a method that can accept two Strings and combine them together.
- Print first String without its first character.
- Then print the second word without its first character.
-
-
-Example;
-            Input:
-                apple
-                banana
-            Output:
-                ppleanana
- */
-}
