+++ src/java_practices/week01/bonusTask/TipCalculator_Melike.java
@@ -1,0 +1,41 @@
+package java_practices.week01.bonusTask;
+
+public class TipCalculator_Melike {
+/*
+    Write a program for a tip calculator.
+    There will be different service quality benchmarks that will determine the tip given.
+    There will also the ability to calculate based on the number of people in the party and if there is a split of the bill or not.
+
+        Poor = 5%, Fair = 10%, Good = 15%, Great = 20%, Excellent = 25%
+
+    The program should ask the user to enter:
+            - Split or No split (Yes or No),
+             - Enter the number of people (number)
+             - Enter the check amount (number)
+             - Service Quality (String)
+
+    And then it should display:
+             Number of people entered
+             Total to pay
+             Total tip
+             Total per person
+             Tip per person:
+
+        Ex:
+            Split or No split (Yes or No)?
+            YeS
+            Enter the number of people
+            4
+            Enter the check amount:
+            476
+            How was the service quality? (Excellent/Great/Good/Fair/Poor)
+            ExCElLeNt
+
+            output:
+                Number of people entered: 4
+                Total to pay: 595.0
+                Total tip: 119.0
+                Total per person: 148.75
+                Tip per person: 29.75
+ */
+}
