+++ src/java_practices/week01/task1/CombineWords_Emir.java
@@ -1,0 +1,21 @@
+package java_practices.week01.task1;
+
+public class CombineWords_Emir {
+
+    public static void main(String[] args) {
+
+
+        System.out.println("Buse");
+    }
+    /*
+    Create a method that can combine two Strings then print the combined version.
+    But if the last letter of the first String and the first letter of the second String is the same, print that character once.
+
+    Example;
+                Input:
+                   String1= one
+                   String2 = eight
+                Output:
+                     oneight
+     */
+}
