+++ src/java_practices/week01/task4/ReallyNeverMind_Melike.java
@@ -1,0 +1,7 @@
+package java_practices.week01.task4;
+
+public class ReallyNeverMind_Melike {
+/*
+Create a method that accept a String.If the String ends with "ly", print "really???" ,  otherwise, print "never mind".
+ */
+}
