+++ src/java_practices/week01/task5/ReplaceX_Buse.java
@@ -1,0 +1,13 @@
+package java_practices.week01.task5;
+
+public class ReplaceX_Buse {
+/*
+Create a method that accept a String.And replace all the x or X with character a.
+
+Example;
+                    Input:
+                        xcodeX
+                    Output:
+                        acodea
+ */
+}
