+++ src/java_practices/week01/task6/StartsWithX_Emir.java
@@ -1,0 +1,16 @@
+package java_practices.week01.task6;
+
+public class StartsWithX_Emir {
+/*
+Create a method that accept a String.
+If the String starts with x, replace it with a.
+
+Example;
+
+                    Input:
+                        xcodex
+                    Output:
+                        acodex
+
+ */
+}
