+++ src/java_practices/week01/task6/StartsWithX_Melike.java
@@ -1,0 +1,16 @@
+package java_practices.week01.task6;
+
+public class StartsWithX_Melike {
+/*
+Create a method that accept a String.
+If the String starts with x, replace it with a.
+
+Example;
+
+                    Input:
+                        xcodex
+                    Output:
+                        acodex
+
+ */
+}
